move stuff from models/data to models
