+++ src/models/channel.rs
