+++ src/models/emoji.rs
