+++ src/models/gateway_events.rs
