+++ src/models/guild_scheduled_event.rs
