+++ src/models/guild_template.rs
