+++ src/models/mod.rs
@@ -1,6 +1,32 @@
 // TODO: jsut make some build script that autogenerates this file
 
-pub mod data;
 pub mod endpoints;
 pub mod locales;
 pub mod types;
+
+pub mod application;
+pub mod audit_log;
+pub mod auto_moderation;
+pub mod channel;
+pub mod directory;
+pub mod discovery;
+pub mod emoji;
+pub mod family_center;
+pub mod gateway_events;
+pub mod guild;
+pub mod guild_scheduled_event;
+pub mod guild_template;
+pub mod integration;
+pub mod invite;
+pub mod message;
+pub mod premium_referral;
+pub mod presence;
+pub mod quest;
+pub mod relationship;
+pub mod soundboard;
+pub mod stage_instance;
+pub mod sticker;
+pub mod team;
+pub mod user;
+pub mod voice;
+pub mod webhook;
