+++ src/models/presence.rs
