+++ src/models/sticker.rs
